Integration tests setup, updated docs
