+++ .travis.yml
@@ -27,6 +27,7 @@ install:
   - pip install --upgrade pip
   - pip install -r requirements/requirements.txt
   - pip install -r requirements/test_requirements.txt
+  - pip install -r requirements/integration_test_requirements.txt
   # work around issues with GPy and GPyOpt setting matplotlib backend
   - "echo 'backend: Agg' > matplotlibrc"
   - pip install .
@@ -34,6 +35,7 @@ install:
 script:
   # - flake8
   - python -m pytest
+  - pytest -c /dev/null integration_tests
 after_success:
   - codecov
 jobs:
