+++ doc/api/emukit.rst
@@ -11,7 +11,6 @@ Subpackages
     emukit.examples
     emukit.experimental_design
     emukit.model_wrappers
-    emukit.models
     emukit.multi_fidelity
     emukit.sensitivity
     emukit.test_functions
