+++ doc/conf.py
@@ -19,7 +19,7 @@ import shutil
 import sys
 from os.path import abspath, dirname
 
-# -- Mock graphing modules - workaround for wrong behavior of some dependencies
+# -- Mock graphing modules - workaround for optinal dependencies as well as for wrong behavior of some required ones
 from unittest.mock import MagicMock
 
 class Mock(MagicMock):
@@ -27,7 +27,7 @@ class Mock(MagicMock):
     def __getattr__(cls, name):
         return MagicMock()
 
-MOCK_MODULES = ['matplotlib', 'pylab', 'matplotlib.pyplot']
+MOCK_MODULES = ['matplotlib', 'pylab', 'matplotlib.pyplot', 'pyrfr.regression', 'pybnn']
 sys.modules.update((module_name, Mock()) for module_name in MOCK_MODULES)
 
 # -- Project information -----------------------------------------------------
