+++ setup.cfg
@@ -8,4 +8,4 @@ addopts =
 [coverage:run]
 branch = true
 source = emukit
-omit = ./tests/*.py, setup.py
+omit = ./tests/*.py, setup.py, ./emukit/examples/*.py
